32. Client Streaming API Server Implementation.
The method doesn't have a request parameter at all.
Instead, it gets a GreetService_LongGreetServer stream,
which the server can use to both read and write messages.
It can receive client messages using its Recv() method and
return its single response using its SendAndClose() method.

The server needs to check the error returned from Recv() after each call.
If this is nil, the stream is still good and it can continue reading;
if it's io.EOF the message stream has ended and the server can return its response.
If it has any other value, we return the error 'as is' so that it'll be translated
to an RPC status by the gRPC layer.