33. Client Streaming API Client Implementation.
The GreetService_LongGreetClient has a Send() method that
we can use to send requests to the server. Once we've finished
writing our client's requests to the stream using Send(),
we need to call CloseAndRecv() on the stream to let gRPC know that
we've finished writing and are expecting to receive a response.

We get our RPC status from the err returned from CloseAndRecv().
If the status is nil, then the first return value from CloseAndRecv()
will be a valid server response.